issue #570: updateFishbowl() fails once

+++ js/client/modules/org/arangodb/aal.js
@@ -482,7 +482,7 @@ function updateFishbowl () {
     filename = undefined;
   }
   catch (err) {
-    if (filename !== undefined) {
+    if (filename !== undefined && fs.exists(filename)) {
       fs.remove(filename);
     }
